DOMA-1500 extracted initialization of `SbbolFintechApi` and removed duplicated code
