+++ apps/condo/domains/organization/integrations/sbbol/sync/syncPaymentRequestsForSubscriptions.js
@@ -1,13 +1,11 @@
 const { ServiceSubscription, ServiceSubscriptionPayment } = require('@condo/domains/subscription/utils/serverSchema')
 const { getSchemaCtx } = require('@core/keystone/schema')
 const dayjs = require('dayjs')
-const { SbbolRequestApi } = require('../SbbolRequestApi')
-const { SbbolFintechApi } = require('../SbbolFintechApi')
+const { initSbbolFintechApi } = require('../SbbolFintechApi')
 const { SBBOL_YEARLY_SUBSCRIPTION_PRICE, SUBSCRIPTION_TYPE, SUBSCRIPTION_PAYMENT_STATUS, SUBSCRIPTION_PAYMENT_CURRENCY } = require('@condo/domains/subscription/constants')
 
 const conf = process.env
 const SBBOL_FINTECH_CONFIG = conf.SBBOL_FINTECH_CONFIG ? JSON.parse(conf.SBBOL_FINTECH_CONFIG) : {}
-const SBBOL_PFX = conf.SBBOL_PFX ? JSON.parse(conf.SBBOL_PFX) : {}
 
 const { logger: baseLogger } = require('../common')
 const { dvSenderFields, BANK_OPERATION_CODE } = require('../constants')
@@ -122,27 +120,8 @@ const postPaymentRequestFor = async (subscription, fintechApi) => {
 const syncPaymentRequestsForSubscriptions = async () => {
     logger.info({ message: 'Start', function: 'syncPaymentRequestsForSubscriptions' })
 
-    let accessToken
-    try {
-        // `service_organization_hashOrgId` is a `userInfo.HashOrgId` from SBBOL, that used to obtain accessToken
-        // for organization, that will be queried in SBBOL using `SbbolFintechApi`.
-        accessToken = await SbbolRequestApi.getOrganizationAccessToken(SBBOL_FINTECH_CONFIG.service_organization_hashOrgId)
-    } catch (e) {
-        logger.error({
-            message: 'Failed to obtain organization access token from SBBOL',
-            error: e.message,
-            hashOrgId: SBBOL_FINTECH_CONFIG.service_organization_hashOrgId,
-        })
-        return
-    }
-
-    const fintechApi = new SbbolFintechApi({
-        accessToken,
-        host: SBBOL_FINTECH_CONFIG.host,
-        port: SBBOL_FINTECH_CONFIG.port,
-        certificate: SBBOL_PFX.certificate,
-        passphrase: SBBOL_PFX.passphrase,
-    })
+    const fintechApi = await initSbbolFintechApi()
+    if (!fintechApi) return
 
     const { keystone: context } = await getSchemaCtx('ServiceSubscription')
     const today = dayjs()
