+++ apps/condo/domains/organization/integrations/sbbol/sync/syncSbbolSubscriptionPaymentRequestsState.js
@@ -1,17 +1,12 @@
 const { logger: baseLogger } = require('../common')
-const { SbbolRequestApi } = require('../SbbolRequestApi')
 const { getSchemaCtx } = require('@core/keystone/schema')
-const { SbbolFintechApi } = require('../SbbolFintechApi')
+const { initSbbolFintechApi } = require('../SbbolFintechApi')
 const { ServiceSubscriptionPayment, ServiceSubscription } = require('@condo/domains/subscription/utils/serverSchema')
 const { SUBSCRIPTION_PAYMENT_STATUS, SUBSCRIPTION_TYPE } = require('@condo/domains/subscription/constants')
 const { executeInSequence } = require('@condo/domains/common/utils/parallel')
 const { SBBOL_PAYMENT_STATUS_MAP, dvSenderFields } = require('../constants')
 const { get } = require('lodash')
 const dayjs = require('dayjs')
-
-const conf = process.env
-const SBBOL_FINTECH_CONFIG = conf.SBBOL_FINTECH_CONFIG ? JSON.parse(conf.SBBOL_FINTECH_CONFIG) : {}
-const SBBOL_PFX = conf.SBBOL_PFX ? JSON.parse(conf.SBBOL_PFX) : {}
 
 const logger = baseLogger.child({ module: 'syncSbbolSubscriptionPaymentRequestsState' })
 
@@ -58,27 +53,8 @@ const syncSbbolSubscriptionPaymentRequestStateFor = async (payment, fintechApi) 
 const syncSbbolSubscriptionPaymentRequestsState = async () => {
     const { keystone: context } = await getSchemaCtx('ServiceSubscriptionPayment')
 
-    let accessToken
-    try {
-        // `service_organization_hashOrgId` is a `userInfo.HashOrgId` from SBBOL, that used to obtain accessToken
-        // for organization, that will be queried in SBBOL using `SbbolFintechApi`.
-        accessToken = await SbbolRequestApi.getOrganizationAccessToken(SBBOL_FINTECH_CONFIG.service_organization_hashOrgId)
-    } catch (e) {
-        logger.error({
-            message: 'Failed to obtain organization access token from SBBOL',
-            error: e.message,
-            hashOrgId: SBBOL_FINTECH_CONFIG.service_organization_hashOrgId,
-        })
-        return
-    }
-
-    const fintechApi = new SbbolFintechApi({
-        accessToken,
-        host: SBBOL_FINTECH_CONFIG.host,
-        port: SBBOL_FINTECH_CONFIG.port,
-        certificate: SBBOL_PFX.certificate,
-        passphrase: SBBOL_PFX.passphrase,
-    })
+    const fintechApi = await initSbbolFintechApi()
+    if (!fintechApi) return
 
     const { CREATED, PROCESSING, STOPPED } = SUBSCRIPTION_PAYMENT_STATUS
