+++ apps/condo/domains/organization/integrations/sbbol/sync/syncSubscriptions.js
@@ -1,5 +1,4 @@
-const { SbbolRequestApi } = require('../SbbolRequestApi')
-const { SbbolFintechApi } = require('../SbbolFintechApi')
+const { initSbbolFintechApi } = require('../SbbolFintechApi')
 const { logger: baseLogger } = require('../common')
 const dayjs = require('dayjs')
 const { ServiceSubscription } = require('@condo/domains/subscription/utils/serverSchema')
@@ -10,7 +9,6 @@ const { getSchemaCtx } = require('@core/keystone/schema')
 
 const conf = process.env
 const SBBOL_FINTECH_CONFIG = conf.SBBOL_FINTECH_CONFIG ? JSON.parse(conf.SBBOL_FINTECH_CONFIG) : {}
-const SBBOL_PFX = conf.SBBOL_PFX ? JSON.parse(conf.SBBOL_PFX) : {}
 
 const logger = baseLogger.child({ module: 'syncSubscriptions' })
 
@@ -111,29 +109,8 @@ const syncSubscriptionsFor = async (advanceAcceptance) => {
 const syncSubscriptions = async (date = null) => {
     if (!date) date = dayjs().format('YYYY-MM-DD')
 
-    logger.info({ message: 'Start syncSubscriptions' })
-
-    let accessToken
-    try {
-        // `service_organization_hashOrgId` is a `userInfo.HashOrgId` from SBBOL, that used to obtain accessToken
-        // for organization, that will be queried in SBBOL using `SbbolFintechApi`.
-        accessToken = await SbbolRequestApi.getOrganizationAccessToken(SBBOL_FINTECH_CONFIG.service_organization_hashOrgId)
-    } catch (e) {
-        logger.error({
-            message: 'Failed to obtain organization access token from SBBOL',
-            error: e.message,
-            hashOrgId: SBBOL_FINTECH_CONFIG.service_organization_hashOrgId,
-        })
-        return
-    }
-
-    const fintechApi = new SbbolFintechApi({
-        accessToken,
-        host: SBBOL_FINTECH_CONFIG.host,
-        port: SBBOL_FINTECH_CONFIG.port,
-        certificate: SBBOL_PFX.certificate,
-        passphrase: SBBOL_PFX.passphrase,
-    })
+    const fintechApi = await initSbbolFintechApi()
+    if (!fintechApi) return
 
     logger.info({ message: 'Checking, whether the user have ServiceSubscription items' })
